pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "vision_transformer"
version = "0.0.1"
description = "This project is an implementation of a vision transformers to solve common tasks in computer vision."
authors = [{ name = "Your name (or your organization/company/team)" }]

readme = "README.md"
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
  "black",
  "flake8",
  "isort",
  "loguru",
  "mkdocs",
  "pip",
  "pytest",
  "python-dotenv",
  "tqdm",
  "typer",
  "requests (>=2.32.3,<3.0.0)",
]
requires-python = "~=3.13.0"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["vision_transformer"]
force_sort_within_sections = true

